collapse into one factory
