+++ pkg/kubectl/cmd/util/BUILD
@@ -5,9 +5,7 @@ go_library(
     srcs = [
         "conversion.go",
         "factory.go",
-        "factory_builder.go",
         "factory_client_access.go",
-        "factory_object_mapping.go",
         "helpers.go",
         "kubectl_match_version.go",
         "printing.go",
