+++ pkg/kubectl/cmd/util/factory_client_access.go
@@ -21,6 +21,7 @@ package util
 import (
 	"fmt"
 	"io"
+	"sync"
 
 	appsv1 "k8s.io/api/apps/v1"
 	appsv1beta1 "k8s.io/api/apps/v1beta1"
@@ -38,47 +39,60 @@ import (
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	restclient "k8s.io/client-go/rest"
+	scaleclient "k8s.io/client-go/scale"
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
+	api "k8s.io/kubernetes/pkg/apis/core"
 	"k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
 	"k8s.io/kubernetes/pkg/kubectl"
+	"k8s.io/kubernetes/pkg/kubectl/cmd/util/openapi"
+	openapivalidation "k8s.io/kubernetes/pkg/kubectl/cmd/util/openapi/validation"
 	"k8s.io/kubernetes/pkg/kubectl/genericclioptions"
 	"k8s.io/kubernetes/pkg/kubectl/genericclioptions/resource"
+	"k8s.io/kubernetes/pkg/kubectl/validation"
 )
 
-type ring0Factory struct {
+type factoryImpl struct {
 	clientGetter genericclioptions.RESTClientGetter
+
+	// openAPIGetter loads and caches openapi specs
+	openAPIGetter openAPIGetter
 }
 
-func NewClientAccessFactory(clientGetter genericclioptions.RESTClientGetter) ClientAccessFactory {
+type openAPIGetter struct {
+	once   sync.Once
+	getter openapi.Getter
+}
+
+func NewFactory(clientGetter genericclioptions.RESTClientGetter) Factory {
 	if clientGetter == nil {
 		panic("attempt to instantiate client_access_factory with nil clientGetter")
 	}
 
-	f := &ring0Factory{
+	f := &factoryImpl{
 		clientGetter: clientGetter,
 	}
 
 	return f
 }
 
-func (f *ring0Factory) ToRESTConfig() (*restclient.Config, error) {
+func (f *factoryImpl) ToRESTConfig() (*restclient.Config, error) {
 	return f.clientGetter.ToRESTConfig()
 }
 
-func (f *ring0Factory) ToRESTMapper() (meta.RESTMapper, error) {
+func (f *factoryImpl) ToRESTMapper() (meta.RESTMapper, error) {
 	return f.clientGetter.ToRESTMapper()
 }
 
-func (f *ring0Factory) ToDiscoveryClient() (discovery.CachedDiscoveryInterface, error) {
+func (f *factoryImpl) ToDiscoveryClient() (discovery.CachedDiscoveryInterface, error) {
 	return f.clientGetter.ToDiscoveryClient()
 }
 
-func (f *ring0Factory) ToRawKubeConfigLoader() clientcmd.ClientConfig {
+func (f *factoryImpl) ToRawKubeConfigLoader() clientcmd.ClientConfig {
 	return f.clientGetter.ToRawKubeConfigLoader()
 }
 
-func (f *ring0Factory) KubernetesClientSet() (*kubernetes.Clientset, error) {
+func (f *factoryImpl) KubernetesClientSet() (*kubernetes.Clientset, error) {
 	clientConfig, err := f.ToRESTConfig()
 	if err != nil {
 		return nil, err
@@ -86,7 +100,7 @@ func (f *ring0Factory) KubernetesClientSet() (*kubernetes.Clientset, error) {
 	return kubernetes.NewForConfig(clientConfig)
 }
 
-func (f *ring0Factory) ClientSet() (internalclientset.Interface, error) {
+func (f *factoryImpl) ClientSet() (internalclientset.Interface, error) {
 	clientConfig, err := f.ToRESTConfig()
 	if err != nil {
 		return nil, err
@@ -94,7 +108,7 @@ func (f *ring0Factory) ClientSet() (internalclientset.Interface, error) {
 	return internalclientset.NewForConfig(clientConfig)
 }
 
-func (f *ring0Factory) DynamicClient() (dynamic.Interface, error) {
+func (f *factoryImpl) DynamicClient() (dynamic.Interface, error) {
 	clientConfig, err := f.ToRESTConfig()
 	if err != nil {
 		return nil, err
@@ -103,11 +117,11 @@ func (f *ring0Factory) DynamicClient() (dynamic.Interface, error) {
 }
 
 // NewBuilder returns a new resource builder for structured api objects.
-func (f *ring0Factory) NewBuilder() *resource.Builder {
+func (f *factoryImpl) NewBuilder() *resource.Builder {
 	return resource.NewBuilder(f.clientGetter)
 }
 
-func (f *ring0Factory) RESTClient() (*restclient.RESTClient, error) {
+func (f *factoryImpl) RESTClient() (*restclient.RESTClient, error) {
 	clientConfig, err := f.ToRESTConfig()
 	if err != nil {
 		return nil, err
@@ -116,8 +130,97 @@ func (f *ring0Factory) RESTClient() (*restclient.RESTClient, error) {
 	return restclient.RESTClientFor(clientConfig)
 }
 
-func (f *ring0Factory) DefaultNamespace() (string, bool, error) {
+func (f *factoryImpl) DefaultNamespace() (string, bool, error) {
 	return f.clientGetter.ToRawKubeConfigLoader().Namespace()
+}
+
+func (f *factoryImpl) ClientForMapping(mapping *meta.RESTMapping) (resource.RESTClient, error) {
+	cfg, err := f.clientGetter.ToRESTConfig()
+	if err != nil {
+		return nil, err
+	}
+	if err := setKubernetesDefaults(cfg); err != nil {
+		return nil, err
+	}
+	gvk := mapping.GroupVersionKind
+	switch gvk.Group {
+	case api.GroupName:
+		cfg.APIPath = "/api"
+	default:
+		cfg.APIPath = "/apis"
+	}
+	gv := gvk.GroupVersion()
+	cfg.GroupVersion = &gv
+	return restclient.RESTClientFor(cfg)
+}
+
+func (f *factoryImpl) UnstructuredClientForMapping(mapping *meta.RESTMapping) (resource.RESTClient, error) {
+	cfg, err := f.clientGetter.ToRESTConfig()
+	if err != nil {
+		return nil, err
+	}
+	if err := restclient.SetKubernetesDefaults(cfg); err != nil {
+		return nil, err
+	}
+	cfg.APIPath = "/apis"
+	if mapping.GroupVersionKind.Group == api.GroupName {
+		cfg.APIPath = "/api"
+	}
+	gv := mapping.GroupVersionKind.GroupVersion()
+	cfg.ContentConfig = resource.UnstructuredPlusDefaultContentConfig()
+	cfg.GroupVersion = &gv
+	return restclient.RESTClientFor(cfg)
+}
+
+func (f *factoryImpl) Validator(validate bool) (validation.Schema, error) {
+	if !validate {
+		return validation.NullSchema{}, nil
+	}
+
+	resources, err := f.OpenAPISchema()
+	if err != nil {
+		return nil, err
+	}
+
+	return validation.ConjunctiveSchema{
+		openapivalidation.NewSchemaValidation(resources),
+		validation.NoDoubleKeySchema{},
+	}, nil
+}
+
+// OpenAPISchema returns metadata and structural information about Kubernetes object definitions.
+func (f *factoryImpl) OpenAPISchema() (openapi.Resources, error) {
+	discovery, err := f.clientGetter.ToDiscoveryClient()
+	if err != nil {
+		return nil, err
+	}
+
+	// Lazily initialize the OpenAPIGetter once
+	f.openAPIGetter.once.Do(func() {
+		// Create the caching OpenAPIGetter
+		f.openAPIGetter.getter = openapi.NewOpenAPIGetter(discovery)
+	})
+
+	// Delegate to the OpenAPIGetter
+	return f.openAPIGetter.getter.Get()
+}
+
+func (f *factoryImpl) ScaleClient() (scaleclient.ScalesGetter, error) {
+	discoClient, err := f.clientGetter.ToDiscoveryClient()
+	if err != nil {
+		return nil, err
+	}
+	restClient, err := f.RESTClient()
+	if err != nil {
+		return nil, err
+	}
+	resolver := scaleclient.NewDiscoveryScaleKindResolver(discoClient)
+	mapper, err := f.clientGetter.ToRESTMapper()
+	if err != nil {
+		return nil, err
+	}
+
+	return scaleclient.New(restClient, mapper, dynamic.LegacyAPIPathResolverFunc, resolver), nil
 }
 
 const (
@@ -326,7 +429,7 @@ func Contains(resourcesList []*metav1.APIResourceList, resource schema.GroupVers
 	return len(resources) != 0
 }
 
-func (f *ring0Factory) Generators(cmdName string) map[string]kubectl.Generator {
+func (f *factoryImpl) Generators(cmdName string) map[string]kubectl.Generator {
 	return DefaultGenerators(cmdName)
 }
